+++ app/services/areas.py
@@ -1,9 +1,15 @@
 from typing import List
-from app.routes.schemas.AreaResponse import AreaResponse, AreaDetailsResponse
-from app.models.models import Polygon, AreaType
-from app.routes.schemas.AreaTypeResponse import AreaTypeResponse
-from app.utils.config import TORTOISE_ORM
 from tortoise import Tortoise
+
+from app.routes.schemas.AreaResponse import AreaResponse, AreaDetailsResponse
+from app.routes.schemas.AreaTypeResponse import AreaTypeResponse
+from app.routes.schemas.PolygonRequest import PolygonRequest
+from app.routes.schemas.PolygonResponse import PolygonIdResponse
+
+from app.models.models import Polygon, AreaType
+from app.persistence.polygon_persistence import get_polygon, create_polygon
+
+from app.utils.config import TORTOISE_ORM
 
 
 async def get_areas_by_type(area_type_id: str) -> List[AreaResponse]:
@@ -44,3 +50,21 @@ async def get_area_details(id: int) -> AreaDetailsResponse | None:
     await Tortoise.close_connections()
 
     return area
+
+
+async def get_or_create_polygon(
+    polygon: PolygonRequest,
+) -> PolygonIdResponse:
+    """
+    Checks if a polygon already exists in the database by comparing its geometry.
+    If it exists, returns the corresponding polygon ID.
+    If it does not exist, creates a new polygon record and returns the new ID.
+    """
+    polygon_geometry = polygon.polygon.geometry
+
+    existing_id = await get_polygon(polygon_geometry)
+    if existing_id is not None:
+        return PolygonIdResponse(polygon_id=existing_id)
+
+    created_id = await create_polygon(polygon_geometry)
+    return PolygonIdResponse(polygon_id=created_id)
